Refactoring part 1 of (hopefully) 2.
Refactored all names to make more sense. I decided I cannot assume the meaning of the given operators because of the ambiguous operators like "&" and ":", so instead of naming trivial operators like "!=" as OPERATOR_NOT_EQUAL_TO, they are now named their literal counterparts: OPERATOR_EXCLAMATION_POINT_EQUALS. I believe this adds clarity to the project, and consistency. The only real leap of faith I am making now is which of the provided special characters are operators and which are delimiters, but that is fairly easy to assume. Only one that was on the fence was ":", but I left it as an operator for now.

Refactored out all literal instances of things for easy reconfiguration (see the do...while loop. I had a nasty bug with this earlier because of it being a literal).

All that should remain is refactoring the logic of the program itself.

Also all of the TODOs (except on logic TODO) are gone, as are most of the relevant warnings CLion has been pushing on me (I WILL NOT USE THE KEYWORD AUTO. It hurts clarity!).

Also removed TokenIdentifier.h by combining it with the Token spec. This makes more sense.

Oh and made all these crazy enum and super specific declarative variables as extern variables to make them globally available.

OH! and made everything const that I possibly could to reduce state issues.